OddOneOut Test
Create a simple design for the app.
Made scripts called 'ChangeScene.cs' to change scenes.
'GamePlay.cs' script contain a list that take in users input form inputfields. the 'List<string>player' has to be static so it can be used for other scenes

Currently the button one scene 'HideAnimal' will change the the text one top called 'PlayNametxt' to player2's name

To do;
when the continue button is click on 'HideAnimal' scene it should switch to 'ShowAnimal' scene to tell the player if they have to animal or if they are the imposter. This should loop until the end of the player list in 'GamePlay.cs'

Need to create a voting page and add up all the vote from the player to see who they think the imposter is. Then show how is the real imposter

Bugs:
when ever you try to rename a player in the player scene it will add a player to the list instead.